+++ fleetfinder/__init__.py
@@ -2,12 +2,7 @@
 Initialize the app
 """
 
-github_url: str = "https://github.com/ppfeufer/aa-fleetfinder"
-
 __title__ = "Fleet Finder"
-__version__ = "0.1.0-alpha.17"
+__version__ = "0.1.0-alpha.18"
 
 __verbose_name__ = "Fleet Finder for Alliance Auth"
-__user_agent_name__ = "Fleet-Finder-for-Alliance-Auth"
-
-__user_agent__ = f"{__verbose_name__} v{__version__} {github_url}"