+++ fleetfinder/constants.py
@@ -5,6 +5,17 @@ Constants used in this app
 # Third Party
 from bravado.exception import HTTPBadGateway, HTTPGatewayTimeout, HTTPServiceUnavailable
 
+# Django
+from django.utils.text import slugify
+
+# AA Fleet Finder
+from fleetfinder import __verbose_name__, __version__
+
+verbose_name_slugified: str = slugify(__verbose_name__, allow_unicode=True)
+github_url: str = "https://github.com/ppfeufer/aa-fleetfinder"
+
+USER_AGENT = f"{verbose_name_slugified} v{__version__} {github_url}"
+
 ESI_ERROR_LIMIT = 50
 ESI_TIMEOUT_ONCE_ERROR_LIMIT_REACHED = 60
 ESI_MAX_RETRIES = 3