+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - CSS moved to bundled HTML template
+- Logger messages changed to f-strings
 - Minimum Requirements
   - Alliance Auth >= 3.0.0
   - Python >= 3.8