+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [In Development] - Unreleased
 
+
+## [0.1.0-alpha.18] - 2022-08-23
+
 ### Added
 
 - `related_name` to `fleet_commander` in model